WIP: Remove unnecessary abstractions
