+++ shared/data/core/src/main/java/com/ivy/data/di/RepositoryBindingsModule.kt
@@ -6,13 +6,6 @@ import com.ivy.data.repository.CurrencyRepository
 import com.ivy.data.repository.ExchangeRatesRepository
 import com.ivy.data.repository.LegalRepository
 import com.ivy.data.repository.TagRepository
-import com.ivy.data.repository.impl.AccountRepositoryImpl
-import com.ivy.data.repository.impl.CategoryRepositoryImpl
-import com.ivy.data.repository.impl.CurrencyRepositoryImpl
-import com.ivy.data.repository.impl.ExchangeRatesRepositoryImpl
-import com.ivy.data.repository.impl.LegalRepositoryImpl
-import com.ivy.data.repository.impl.TagRepositoryImpl
-import com.ivy.data.repository.TransactionRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -22,20 +15,20 @@ import dagger.hilt.components.SingletonComponent
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryBindingsModule {
     @Binds
-    abstract fun bindAccountRepo(repo: AccountRepositoryImpl): AccountRepository
+    abstract fun bindAccountRepo(repo: AccountRepository): AccountRepository
 
     @Binds
-    abstract fun bindCategoryRepo(repo: CategoryRepositoryImpl): CategoryRepository
+    abstract fun bindCategoryRepo(repo: CategoryRepository): CategoryRepository
 
     @Binds
-    abstract fun bindExchangeRatesRepo(repo: ExchangeRatesRepositoryImpl): ExchangeRatesRepository
+    abstract fun bindExchangeRatesRepo(repo: ExchangeRatesRepository): ExchangeRatesRepository
 
     @Binds
-    abstract fun bindTagsRepo(repo: TagRepositoryImpl): TagRepository
+    abstract fun bindTagsRepo(repo: TagRepository): TagRepository
 
     @Binds
-    abstract fun bindCurrencyRepo(repo: CurrencyRepositoryImpl): CurrencyRepository
+    abstract fun bindCurrencyRepo(repo: CurrencyRepository): CurrencyRepository
 
     @Binds
-    abstract fun bindLegalRepo(repo: LegalRepositoryImpl): LegalRepository
+    abstract fun bindLegalRepo(repo: LegalRepository): LegalRepository
 }
