+++ shared/data/core/src/main/java/com/ivy/data/repository/LegalRepository.kt
@@ -1,6 +1,21 @@
 package com.ivy.data.repository
 
-interface LegalRepository {
-    suspend fun isDisclaimerAccepted(): Boolean
-    suspend fun setDisclaimerAccepted(accepted: Boolean)
+import com.ivy.base.threading.DispatchersProvider
+import com.ivy.data.datasource.LocalLegalDataSource
+import kotlinx.coroutines.withContext
+import javax.inject.Inject
+
+class LegalRepository @Inject constructor(
+    private val localLegalDataSource: LocalLegalDataSource,
+    private val dispatchers: DispatchersProvider
+) {
+    suspend fun isDisclaimerAccepted(): Boolean = withContext(dispatchers.io) {
+        localLegalDataSource.getIsDisclaimerAccepted() ?: false
+    }
+
+    suspend fun setDisclaimerAccepted(
+        accepted: Boolean
+    ): Unit = withContext(dispatchers.io) {
+        localLegalDataSource.setDisclaimerAccepted(accepted)
+    }
 }
