+++ shared/data/core/src/main/java/com/ivy/data/repository/fake/FakeCurrencyRepository.kt
@@ -3,7 +3,6 @@ package com.ivy.data.repository.fake
 import com.ivy.data.db.dao.read.SettingsDao
 import com.ivy.data.db.dao.write.WriteSettingsDao
 import com.ivy.data.repository.CurrencyRepository
-import com.ivy.data.repository.impl.CurrencyRepositoryImpl
 import com.ivy.base.TestDispatchersProvider
 import org.jetbrains.annotations.VisibleForTesting
 
@@ -11,7 +10,7 @@ import org.jetbrains.annotations.VisibleForTesting
 class FakeCurrencyRepository(
     settingsDao: SettingsDao,
     writeSettingsDao: WriteSettingsDao,
-    private val currencyRepository: CurrencyRepository = CurrencyRepositoryImpl(
+    private val currencyRepository: CurrencyRepository = CurrencyRepository(
         settingsDao = settingsDao,
         writeSettingsDao = writeSettingsDao,
         dispatchersProvider = TestDispatchersProvider,
