+++ shared/data/core/src/test/java/com/ivy/data/repository/AccountRepositoryTest.kt
@@ -1,4 +1,4 @@
-package com.ivy.data.repository.impl
+package com.ivy.data.repository
 
 import com.ivy.base.TestDispatchersProvider
 import com.ivy.data.DataObserver
@@ -11,9 +11,8 @@ import com.ivy.data.model.AccountId
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.model.primitive.ColorInt
 import com.ivy.data.model.primitive.NotBlankTrimmedString
-import com.ivy.data.repository.AccountRepository
 import com.ivy.data.repository.fake.FakeCurrencyRepository
-import com.ivy.data.repository.fake.fakeRepositoryMakeFactory
+import com.ivy.data.repository.fake.fakeRepositoryMemoFactory
 import com.ivy.data.repository.mapper.AccountMapper
 import io.kotest.matchers.shouldBe
 import io.mockk.coEvery
@@ -24,10 +23,9 @@ import io.mockk.runs
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
-import java.time.Instant
 import java.util.UUID
 
-class AccountRepositoryImplTest {
+class AccountRepositoryTest {
     val accountDao = mockk<AccountDao>()
     val writeAccountDao = mockk<WriteAccountDao>()
     val writeEventBus = mockk<DataObserver>(relaxed = true)
@@ -37,12 +35,12 @@ class AccountRepositoryImplTest {
     @Before
     fun setup() {
         val settingsDao = FakeSettingsDao()
-        repository = AccountRepositoryImpl(
+        repository = AccountRepository(
             mapper = AccountMapper(FakeCurrencyRepository(settingsDao, settingsDao)),
             accountDao = accountDao,
             writeAccountDao = writeAccountDao,
             dispatchersProvider = TestDispatchersProvider,
-            memoFactory = fakeRepositoryMakeFactory(),
+            memoFactory = fakeRepositoryMemoFactory(),
         )
     }
 
@@ -87,8 +85,6 @@ class AccountRepositoryImplTest {
             icon = null,
             includeInBalance = true,
             orderNum = 1.0,
-            lastUpdated = Instant.EPOCH,
-            removed = false
         )
     }
 
@@ -170,8 +166,6 @@ class AccountRepositoryImplTest {
                 icon = null,
                 includeInBalance = true,
                 orderNum = 1.0,
-                lastUpdated = Instant.EPOCH,
-                removed = false
             ),
             Account(
                 id = account2Id,
@@ -181,8 +175,6 @@ class AccountRepositoryImplTest {
                 icon = null,
                 includeInBalance = true,
                 orderNum = 2.0,
-                lastUpdated = Instant.EPOCH,
-                removed = false
             )
         )
     }
@@ -230,8 +222,6 @@ class AccountRepositoryImplTest {
                 icon = null,
                 includeInBalance = true,
                 orderNum = 1.0,
-                lastUpdated = Instant.EPOCH,
-                removed = false
             )
         )
     }
@@ -273,8 +263,6 @@ class AccountRepositoryImplTest {
             icon = null,
             includeInBalance = true,
             orderNum = 1.0,
-            lastUpdated = Instant.EPOCH,
-            removed = false
         )
 
         // when
@@ -313,8 +301,6 @@ class AccountRepositoryImplTest {
                 icon = null,
                 includeInBalance = true,
                 orderNum = 1.0,
-                lastUpdated = Instant.EPOCH,
-                removed = false
             ),
             Account(
                 id = account2Id,
@@ -324,8 +310,6 @@ class AccountRepositoryImplTest {
                 icon = null,
                 includeInBalance = true,
                 orderNum = 2.0,
-                lastUpdated = Instant.EPOCH,
-                removed = false
             )
         )
