+++ shared/domain/src/androidTest/java/com/ivy/domain/usecase/SyncExchangeRatesUseCaseTest.kt
@@ -11,7 +11,6 @@ import com.ivy.data.di.KtorClientModule
 import com.ivy.data.model.primitive.AssetCode
 import com.ivy.data.remote.impl.RemoteExchangeRatesDataSourceImpl
 import com.ivy.data.repository.ExchangeRatesRepository
-import com.ivy.data.repository.impl.ExchangeRatesRepositoryImpl
 import com.ivy.data.repository.mapper.ExchangeRateMapper
 import com.ivy.domain.usecase.exchange.SyncExchangeRatesUseCase
 import io.kotest.assertions.arrow.core.shouldBeRight
@@ -36,7 +35,7 @@ class SyncExchangeRatesUseCaseTest {
         val context = ApplicationProvider.getApplicationContext<Context>()
         db = Room.inMemoryDatabaseBuilder(context, IvyRoomDatabase::class.java).build()
 
-        repository = ExchangeRatesRepositoryImpl(
+        repository = ExchangeRatesRepository(
             exchangeRatesDao = db.exchangeRatesDao,
             writeExchangeRatesDao = db.writeExchangeRatesDao,
             mapper = ExchangeRateMapper(),
